Calculate specimen density

Given a wet mass (wet_mass) and buoyant mass (buoy_mass), this function calculates the density of a specimen. Bouyant mass is the downward force exerted while the specimen is in water, typically measured by having a measuring plate within a tank of water suspended from a balance above the tank. For species such as molluscs this can be used to determine the ratio of shell to tissue mass. It can also be used to calculate a specimen volume for correcting the water volume of a respirometer (see eff_vol). The temperature (t) and salinity (S) at which the buoyant mass was determined are required (as is atmospheric pressure (P) which defaults to 1.013253 bar, if not otherwise specified).

spec_density(wet_mass = NULL, buoy_mass = NULL, t = NULL, S = NULL, P
  = 1.013253)

Arguments

wet_mass
numeric (kg). Wet mass of specimen.

buoy_mass
numeric (kg). Buoyant mass of specimen.

t
numeric (°C). Water temperature at which buoy_mass was determined.

S
numeric (ppt). Water salinity at which buoy_mass was determined.

P
numeric (bar). Atmospheric pressure at which buoy_mass was determined. Defaults to 1.013253.

Inputs

Inputs must be in SI units:

wet_mass and buoy_mass in kg

t in °C

S in ppt

P in bar (Defaults to 1.013253)

Positively buoyant specimens

The function should work with positively buoyant specimens (i.e. less dense than water). Given a bouyant weight measured as buoyancy (i.e. lift force on a balance while in water), this can be entered as a negative value for buoy_mass and the function should return the correct density value. However, none of this has been tested.

Output

Output is a single numeric value for the density in kg/m^3.

See also